+++ actionpack/test/dispatch/system_testing/driver_test.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "abstract_unit"
+require "support/system_helper"
 require "action_dispatch/system_testing/driver"
 require "selenium/webdriver"
