+++ actionpack/test/dispatch/system_testing/screenshot_helper_test.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "abstract_unit"
+require "support/system_helper"
 require "action_dispatch/system_testing/test_helpers/screenshot_helper"
 require "capybara/dsl"
 require "selenium/webdriver"
