+++ actionpack/test/dispatch/system_testing/system_test_case_test.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "abstract_unit"
+require "support/system_helper"
 require "selenium/webdriver"
 
 class SetDriverToRackTestTest < DrivenByRackTest
